requirements.txt
asgiref==3.9.1
attrs==25.3.0
boto3==1.34.162
botocore==1.34.162
django==5.2.6
django-cors-headers==4.7.0
django-environ==0.12.0
django-storages==1.14.6
djangorestframework==3.16.1
drf-nested-routers==0.94.2
drf-spectacular==0.28.0
gunicorn==23.0.0
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
packaging==25.0
pillow==11.3.0
psycopg==3.2.9
psycopg-binary==3.2.9
python-dateutil==2.9.0.post0
pyyaml==6.0.2
referencing==0.36.2
rpds-py==0.27.1
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
